MDED-58 Добавлен кастинг до IVisirer
